==> 01 IMBD Project.ipynb <==
# IMBD Project 03 (Core) 
04/22/2020

## Part 01 Data Clean up

IMDB Provides Several Files with varied information for Movies, TV Shows, Made for TV Movies, etc.

Data Dictionary: https://www.imdb.com/interfaces/

Downloads page: https://datasets.imdbws.com/

### Links for data file

title basics - https://datasets.imdbws.com/title.basics.tsv.gz

title ratings - https://datasets.imdbws.com/title.ratings.tsv.gz

title akas - https://datasets.imdbws.com/title.akas.tsv.gz

import pandas as pd
import numpy as np

basics_url = "https://datasets.imdbws.com/title.basics.tsv.gz"
ratings_url = "https://datasets.imdbws.com/title.ratings.tsv.gz"
akas_url = "https://datasets.imdbws.com/title.akas.tsv.gz"

# load data
basics_df = pd.read_csv(basics_url, sep='\t', low_memory=False)
ratings_df = pd.read_csv(ratings_url, sep='\t', low_memory=False)
akas_df = pd.read_csv(akas_url, sep='\t', low_memory=False)


basics_df.head()

ratings_df.head()

akas_df.head()

Preposing for Title basics

#replace \n with np.nan 
basics_df = basics_df.replace({'\\N':np.nan})

basics_df.head()

#eliminate movies that are Nan for runtimeMinutes and genres
basics_df = basics_df.dropna(subset=['runtimeMinutes', 'genres', 'startYear'])

#only those with tile movie
basics_df = basics_df.loc[basics['titleType']=='movie']

there are nan in startYear so going to drop nan values above

#only movies with start year 2000-2022
basics_df['startYear'] = basics_df['startYear'].astype(int)
basics_df= basics.loc[(basics_df['startYear'] >=2000) & (basics_df['startYear'] <=2022)]

#Eliminate documentaries
is_documentary = basics_df['genres'].str.contains('documentary', case=False)
basics_df = basics_df[~is_documentary]

basics_df.head()

### Preposing for AKAs

keep only US entries.

Replace "\N" with np.nan

akas_df = akas_df.replace({'\\N':np.nan})

akas_df.head()

#US entries only

akas_df = akas_df.loc[akas_df['region']=='US']


### Preposing for Ratings

ratings_df = ratings_df.replace({'\\N':np.nan})

ratings_df.head()

Filtering one data frame based on another

keepers = basics_df['tconst'].isin(akas_df['titleId'])
basics_df = basics_df[keepers]
basics_df

## Creating a "Data" folder for saving files

# example making new folder with os
import os
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")

#df to files

basics_df.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
akas_df.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)
ratings_df.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

#open saved file and preview again
basics = pd.read_csv('Data/title_basics.csv.gz', low_memory=False)
basics.head()

#open saved file and preview again
akas = pd.read_csv('Data/title_akas.csv.gz', low_memory=False)
akas.head()

#open saved file and preview again
ratings = pd.read_csv('Data/title_ratings.csv.gz', low_memory=False)
ratings.head()
