==> 02 TMDB Project 3 part 2 ( extraction with API).ipynb <==
# Business Problem


For this project, you have been hired to produce a MySQL database on Movies from a subset of IMDB's publicly available dataset. Ultimately, you will use this database to analyze what makes a movie successful, and will provide recommendations to the stakeholder on how to make a successful movie.

## Your Stakeholder Wants More Data!
After investigating the preview of your data from Part 1, your stakeholder realized that there is no financial information included in the IMDB data (e.g. budget or revenue).

This will be a major roadblock when attempting to analyze which movies are successful and must be addressed before you will be able to determine which movies are successful.
Your stakeholder identified The Movie Database (TMDB) as a great source of financial data 

(https://www.themoviedb.org/). Thankfully, TMDB offers a free API for programmatic access to their data!


Your stakeholder wants you to extract the budget, revenue, and MPAA Rating (G/PG/PG-13/R), which is also called "Certification".

# Standard Imports
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# Additional Imports
import os, json, math, time

from tqdm.notebook import tqdm_notebook

# Load API Credentials

with open('/Users/vinosraj/.secret/tmdb_api.json') as f: #default 'r' default so dont need to add
    login = json.load(f)
login.keys()

# Instantiate tmdbAPI Variable
import tmdbsimple as tmdb
tmdb.API_KEY =  login['api-key']

https://developers.themoviedb.org/3/movies/get-movie-details

for API details

## Specify fodler for saving data
FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok=True)

def get_movie_with_rating(movie_id):
    #get move and release date
    movie=tmdb.Movies(movie_id)
    # construct output dict
    movie_info = movie.info()
    release = movie.releases()
    for c in release ['countries']:
        if c['iso_3166_1'] =='US': #else function &*
            movie_info['certification'] = c['certification']
        else:
            movie_info['certification'] = 'Not Available'
    return movie_info

def write_json(new_data, filename): 
    """Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""    
    
    with open(filename,'r+') as file:
        # First we load existing data into a dict.
        file_data = json.load(file)
        ## Choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
             file_data.append(new_data)
        # Sets file's current position at offset.
        file.seek(0)
        # convert back to json.
        json.dump(file_data, file)

basics_file = "Data/title_basics.csv.gz"

basics = pd.read_csv(basics_file, low_memory=False)

basics.head()

# set our API call parameters 

YEARS_TO_GET = [2000,2001]


# Start of OUTER loop
for YEAR in tqdm_notebook (YEARS_TO_GET,desc='YEARS',position=0):
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'

    ## Check if JSON_FILE exists
    file_exists = os.path.isfile(JSON_FILE)

    ## If it does not exist: 
    if file_exists == False:

        ## CREATE ANY NEEDED FOLDERS
        # Get the Folder Name only
        folder = os.path.dirname(JSON_FILE)

         ## If JSON_FILE included a folder:
        if len(folder)>0:
            # create the folder
            os.makedirs(folder,exist_ok=True)


        ## INFORM USER AND SAVE EMPTY LIST
        print(f"[i] {JSON_FILE} not found. Saving empty list to file.")


        ## save the first page of results
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)  
    ## If it exists, inform user
    else:
        print(f"[i] {JSON_FILE} already exists.")

    #Saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()
    # saving movie ids to list
    movie_ids = df['tconst'].copy()#.to_list()

    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    previous_df

    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]

#Get index and movie id from list
# INNER Loop
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                      desc=f'Movies from {YEAR}',
                                      position=1,
                                      leave=True):
            # Attempt to retrieve then data for the movie id
        try:
            temp = get_movie_with_rating(movie_id)  #This uses your pre-made function
           # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)

            # If it fails,  make a dict with just the id and None for certification.
        except Exception as e:
                continue

final_2000_df = pd.read_json('Data/tmdb_api_results_2000.json')
final_2000_df.to_csv(f"{FOLDER}final_tmdb_data_2000.csv.gz", compression="gzip", index=False)

final_2001_df = pd.read_json('Data/tmdb_api_results_2001.json')
final_2001_df.to_csv(f"{FOLDER}final_tmdb_data_2001.csv.gz", compression="gzip", index=False)

#import glob
#glob.glob ("Data/*.json")
