==> 03 IMDB Creating SQL Database.ipynb <==
# My SQL Preparation for IMDB Project

## Required Transformation steps:
 
#### Normalize Genre:

Convert the single string of genres from title  basics into 2 new tables.

    1. title_genres: with the columns:
            tconst
            genre_id
    2. genres:
            genre_id
            genre_name
            
#### Discard unnecessary information:

For the title basics table, drop the following columns:

- "original_title" (we will use the primary title column instead)
- "isAdult" ("Adult" will show up in the genres so this is redundant information).
- "titleType" (every row will be a movie).
- "genres" and other variants of genre (genre is now represented in the 2 new tables described above.

Do not include the title_akas table in your SQL database.

- You have already filtered out the desired movies using this table and the remaining data is mostly nulls and not of-interest to the stakeholder.

#import necessary libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import json
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import create_database, database_exists

username = 'root'
password ='Sachvin2017'



connection = f'mysql+pymysql://{username}:{password}@localhost/Movies'
engine = create_engine(connection)

## Check if database exists, if not, create it
if database_exists(connection) == False: create_database(connection)
else: print('The database already exists.')

title_basics_df = pd.read_csv('Data/title_basics.csv.gz')


title_basics_df.head()


Splitting columns in Genres

title_basics_df['genres_split'] = title_basics_df['genres'].str.split(',')
title_basics_df

#Explode to sepreate the list of genres to seperate new rows
exploded_genres_df = title_basics_df.explode('genres_split')
exploded_genres_df

unique_genres = sorted(exploded_genres_df['genres_split'].unique())
unique_genres


## Table Creation

New table #title_genre

title_genres = exploded_genres_df[['tconst', 'genres_split']].copy()
title_genres.head()

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## make new integer genre_id and drop string genres
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns='genres_split')
title_genres

#manually make dataframe with named cols from the .keyd and .values
genre = pd.DataFrame({'genre_name': genre_map.keys(),
                            'genre_id': genre_map.values()})
genre.head()

## SQL Table creation

### Title_basics table

title_basics_df.head()

title_basics = title_basics_df.drop(columns = ['titleType', 'originalTitle',
                                                  'isAdult', 'endYear', 'genres', 'genres_split'])

title_basics.info()

title_basics.dtypes

# Calculate max string lengths for object columns ('tconst' and 'primaryTitle')
key_len = title_basics_df['tconst'].fillna('').map(len).max()
title_len = title_basics_df['primaryTitle'].fillna('').map(len).max()

# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    'tconst': String(key_len+1), 
    'primaryTitle': Text(title_len+1),
    'startYear': Integer(),
    'runtimeMinutes': Integer()}

title_basics.to_sql('title_basics',engine,dtype=df_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')


q = """
SELECT * 
FROM title_basics LIMIT 3
;"""
pd.read_sql(q, engine)

### title_genres table

title_genres.head()

title_genres.info()

## Calculate max string lengths for object columns
key_len = title_genres['tconst'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    'genre_id': Integer()}

#save to sql with dtype and index=false
title_genres.to_sql('title_genres', engine, dtype=df_schema, 
                   if_exists = 'replace', index=False)

q = """
SELECT * 
FROM title_genres LIMIT 3
;"""
pd.read_sql(q, engine)

### Genre Table

genre.info()

## Calculate max string lengths for object columns
key_len = genre['genre_name'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    'genre_id': Integer(),
    "genre_name": String(key_len+1) 
    }

#save to sql with dtype and index=false
genre.to_sql('genres', engine, dtype=df_schema,
                   if_exists = 'replace', index=False)

# Use the sqlalchemy engine to update the table and set genre_id as the primary key
engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);')

q = """
SELECT * 
FROM genres LIMIT 3
;"""
pd.read_sql(q, engine)

### Title_ Rating table

title_ratings_df = pd.read_csv('Data/title_ratings.csv.gz')
title_ratings_df.head()


title_ratings_df.info()


# Calculate max string lengths for object columns
key_len = title_ratings_df['tconst'].fillna('').map(len).max()

# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    'averageRating':Float(),
    'numVotes':Integer()}

# Save to sql with dtype and index=False
title_ratings_df.to_sql('title_ratings',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

# Use the sqlalchemy engine to update the table and set tconst as the primary key
engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

q = """
SELECT * 
FROM title_ratings LIMIT 3
;"""
pd.read_sql(q, engine)

### tmdb_data

tmdb_df = pd.read_csv('Data/tmdb_results_combined.csv.gz')
tmdb_df.head()

tmdb_df.info()


tmdb_df = tmdb_df.drop(columns = ['adult', 'backdrop_path','belongs_to_collection',
                               'genres', 'homepage', 'id', 'original_language', 
                              'original_title', 'overview', 'popularity', 
                              'poster_path', 'production_companies',
                               'production_countries', 'release_date', 
                              'runtime', 'spoken_languages', 'status', 
                              'tagline', 'title', 'video', 'vote_average', 
                              'vote_count'])
tmdb_df.info()

tmdb_df = tmdb_df.iloc[1: , :]

# Calculate max string lengths for object columns ('tconst' and 'primaryTitle')
imdb_id_len = tmdb_df['imdb_id'].fillna('').map(len).max()

certification_len = tmdb_df['certification'].fillna('').map(len).max()



# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    'imdb_id': String(imdb_id_len+1), 
    'budget':Float(), 
    'revenue':Float(),
    'certification': String(certification_len+1)}

tmdb_df.to_sql('tmdb_data', engine, dtype=df_schema, 
              if_exists='replace', index=False)


# Use the sqlalchemy engine to update the table and set imdb_id as the primary key
engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

q = """
SELECT * 
FROM tmdb_data LIMIT 3
;"""
pd.read_sql(q, engine)

## All SQL tables

q = """SHOW TABLES;"""
pd.read_sql(q, engine)
